feat: migrate database from PostgreSQL to MySQL with UUID support
- Updated .env.example for MySQL configuration.
- Changed ID types from int to string (UUID) in Category and Product entities, repositories, and use cases.
- Modified Eloquent models to use UUIDs as primary keys.
- Updated API documentation to reflect UUID changes for categories and products.
- Adjusted validation rules for category_id to expect UUIDs.
- Enhanced error handling for invalid UUID inputs.
- Updated tests to accommodate UUIDs for category and product IDs.

+++ app/Presentation/Products/Resources/ProductResource.php
@@ -15,7 +15,7 @@ class ProductResource extends JsonResource
             'description' => $this->description,
             'price' => (float) $this->price,
             'stock' => (int) $this->stock,
-            'category_id' => (int) ($this->categoryId ?? $this->category_id),
+            'category_id' => $this->categoryId ?? $this->category_id,
             'created_at' => $this->createdAt ?? $this->created_at,
             'updated_at' => $this->updatedAt ?? $this->updated_at,
         ];
